--- a/bind.js
+++ b/bind.js
@@ -31,8 +31,9 @@
 				continue;
 			}
 
-			if (typeof item !== "object")
+			if (typeof item !== "object") {
 				continue;
+			}
 
 			if (isArray(item)) {
 				if ((item = classNames.apply(this, item))){