--- a/index.js
+++ b/index.js
@@ -14,29 +14,23 @@
 	function classNames() {
 		var len = arguments.length;
 
-		if (!len){
-			return "";
-		}
+		if (!len) return '';
 
-		var str = "", item, i, n;
+		var str = '', item, i, n;
 		for (i = 0; i < len; i++) {
-			if (!(item = arguments[i])) {
-				continue;
-			}
+			if (!(item = arguments[i]))	continue;
 
-			if (typeof item === "string" || typeof item === "number") {
-				str && (str += " ");
+			if (typeof item === 'string' || typeof item === 'number') {
+				str && (str += ' ');
 				(str += item);
 				continue;
 			}
 
-			if (typeof item !== "object") {
-				continue;
-			}
+			if (typeof item !== 'object') continue;
 
 			if (isArray(item)) {
 				if ((item = classNames.apply(this, item))){
-					str && (str += " ");
+					str && (str += ' ');
 					(str += item);
 				}
 
@@ -45,7 +39,7 @@
 
 			for (n in item) {
 				if (hasOwnProperty.call(item, n) && item[n] && n){
-					str && (str += " ");
+					str && (str += ' ');
 					(str += n);
 				}
 			}