@JedWatson 2019-07-05
Those are some big improvements to the benchmarks @xobotyi, awesome!

I'm reviewing this now and can't see any issues with merging your changes.

Out of curiosity, would you mind annotating the PR (adding your own review is a good way to do this) with some explanations of why the changes you made improved the performance?

Thanks!