@newyork-anthonyng 2019-07-05
The original implementation did something similar. 
This [PR](https://github.com/JedWatson/classnames/pull/50#issuecomment-144703425) changed it to the current implementation with arrays. It seemed like _readability_ was the main reason of switching over.