@xobotyi 2019-07-06
@newyork-anthonyng looks kinda weird to me, to speculate readability of code you'll never read with obvious performance  tradeoff :japanese_goblin: