@krawaller 2019-08-27
It very likely won't make a measurable difference, but I'm curious if string constants would be an improvement.

```javascript
var EMPTY = ''
var SPACE = ' ' // (except regex already called this)
var OBJECT = 'object'
var FUNCTION = 'function'
// ...etc
```

It might well be that variables perform worse than literals (when not using `const`) since they are mutable and literals aren't.