@xobotyi 2019-08-27
@krawaller there are no perf improvements if to store them in variables, even more - it will slow down the code due to disabling the typecheck engine optimisation;
```js
typeof a === 'string'
``` 
is always faster than 
```js
var STR = 'string';
typeof a === STR;
```
or even
```js
var type = typeof a;
type === 'string'
```